Add DynamicSelector for dynamic selection method configuration
Introduced a new `DynamicSelector` object in the `dynamic_selector.rs` file and modified the `main.rs` file to use `DynamicSelector`. The `DynamicSelector` allows for dynamic configuration of the selection method used in the genetic algorithm (Maximize, Roulette, Tournament). The `run` function was also updated accordingly to incorporate the `DynamicSelector` as an argument. Commented out some lines related to logging for cleaning purposes. This change makes it more flexible to switch between different selection methods in the genetic algorithm.